ShipRouteIECConverter Operator Conversions now possible for both directions
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@21097 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 569e86e024bf07fea0cc66f153c08470087b9bd9

+++ server/wrapper/shiproutes/src/de/uniol/inf/is/odysseus/wrapper/shiproutes/json/element/route/Waypoint.java
@@ -39,7 +39,7 @@ public class Waypoint implements IShipRouteElement {
 	private Integer crosstracklimit_stbd_m;
 	private Integer crosstracklimit_port_m;
 	private Double speed_kts;
-	private Integer ETA;
+	private Long ETA;
 	private Double speedlimit_upper_kts;
 	private Double speedlimit_lower_kts;
 	private Double distance_to_wp_nm;
@@ -181,11 +181,11 @@ public class Waypoint implements IShipRouteElement {
 		this.speed_kts = speed_kts;
 	}
 
-	public Integer getETA() {
+	public Long getETA() {
 		return ETA;
 	}
 
-	public void setETA(Integer eTA) {
+	public void setETA(Long eTA) {
 		ETA = eTA;
 	}
